code/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

adjustText==1.3.0
datasets==3.2.0
einops==0.8.0
h5py==3.12.1
ipython==8.26.0
matplotlib==3.9.2
nibabel==5.3.2
numpy==1.21.5
pandas==2.2.2
Pillow==9.0.1
pycocotools==2.0+nv0.8.0
pycortex==1.2.11
pymp==0.0.6
scikit-learn==1.5.1
scipy==1.14.0
seaborn==0.13.2
statsmodels==0.14.4
timm==1.0.13
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.67.0
transformers==4.48.0
umap==0.1.1
